Добавлена проверка текущего пользователя с владельцем списка, который нужно изменить/удалить.

+++ src/main/java/ru/ahmed/anam/todoservice/controllers/TodoListController.java
@@ -6,7 +6,6 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import ru.ahmed.anam.todoservice.domain.TodoElement;
 import ru.ahmed.anam.todoservice.domain.TodoList;
@@ -18,6 +17,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
+
+import static java.lang.String.format;
 
 @Log
 @RestController
@@ -35,26 +36,39 @@ public class TodoListController {
     }
 
     @GetMapping
-    public Iterable<TodoList> getAllLists(){
+    public Iterable<TodoList> getAllLists() {
         final User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         return listRepository.findAllByUserUsername(user.getUsername());
     }
 
     @GetMapping("/{listId}")
-    public ResponseEntity<TodoList> getListById(@PathVariable("listId") Long listId){
+    public ResponseEntity<TodoList> getListById(@PathVariable("listId") Long listId) {
+        final User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         final Optional<TodoList> optResult = listRepository.findById(listId);
-        return optResult
-                .map(
-                        todoList -> new ResponseEntity<>(todoList, HttpStatus.OK)
-                )
-                .orElseGet(
-                        () -> new ResponseEntity<>(null, HttpStatus.NOT_FOUND)
+        if (optResult.isPresent()) {
+            final TodoList todoList = optResult.get();
+            final User listOwner = todoList.getUser();
+            if (currentUser.equals(listOwner)) {
+                //возвращаем туду-лист только если он принадлежит текущему пользователю
+                return new ResponseEntity<>(todoList, HttpStatus.OK);
+            } else {
+                log.log(
+                        Level.WARNING,
+                        format(
+                                "Пользователь '%s' пытался получить список (listId=%d), принадлежащий другому другому пользователю - '%s'",
+                                currentUser.getUsername(), listId, listOwner.getUsername()
+                        )
                 );
+                return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            }
+        } else {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping(consumes = "application/json")
     @ResponseStatus(HttpStatus.CREATED)
-    public TodoList postTodoList(@RequestBody TodoList todoList){
+    public TodoList postTodoList(@RequestBody TodoList todoList) {
         final User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         todoList.setUser(user);
         List<TodoElement> todoListElements = todoList.getElements();
@@ -67,21 +81,57 @@ public class TodoListController {
 
     @DeleteMapping(path = "/{listId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTodoList(@PathVariable("listId") Long listId){
+    public void deleteTodoList(@PathVariable("listId") Long listId) {
+        final User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        final Optional<TodoList> list = listRepository.findById(listId);
+        if (list.isEmpty()) {
+            log.log(Level.WARNING, format("Список {id=%d} не найден", listId));
+            return;
+        }
+        final TodoList todoList = list.get();
+        final User listUser = todoList.getUser();
+        if (!listUser.equals(currentUser)) {
+            log.log(
+                    Level.WARNING,
+                    format(
+                            "Пользователь '%s' пытался удалить список {id=%d} (принадлежащий пользователю '%s')",
+                            currentUser.getUsername(),
+                            listId,
+                            listUser.getUsername()
+                    )
+            );
+            return;
+        }
         try {
             listRepository.deleteById(listId);
-        } catch (EmptyResultDataAccessException e){
+        } catch (EmptyResultDataAccessException e) {
             log.log(Level.WARNING, String.format("Error occured during deleting a todo-list with ID={%d}", listId), e);
         }
     }
 
     @PatchMapping(path = "/element/{elementId}/done")
-    public ResponseEntity<TodoElement> markElementAsDone(@PathVariable("elementId") Long elementId){
+    public ResponseEntity<TodoElement> markElementAsDone(@PathVariable("elementId") Long elementId) {
+        final User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         final Optional<TodoElement> optTodoElement = elementRepository.findById(elementId);
-        if(optTodoElement.isPresent()){
-            final TodoElement updatedTodoElement = optTodoElement.get().done();
-            elementRepository.save(updatedTodoElement);
-            return new ResponseEntity<>(updatedTodoElement, HttpStatus.OK);
+        if (optTodoElement.isPresent()) {
+            final TodoElement todoElement = optTodoElement.get();
+            final TodoList todoList = listRepository.findByElementsContains(todoElement);
+            final User listOwner = todoList.getUser();
+            if (listOwner.equals(currentUser)) {
+                //помечаем элемент списка только если лист, в состав которого он входит, принадлежит текущему пользователю
+                final TodoElement updatedTodoElement = todoElement.done();
+                elementRepository.save(updatedTodoElement);
+                return new ResponseEntity<>(updatedTodoElement, HttpStatus.OK);
+            } else {
+                log.log(
+                        Level.WARNING,
+                        format(
+                                "Пользователь '%s' пытался изменить элемент списка, принадлежащий другому другому пользователю - '%s'",
+                                currentUser.getUsername(), listOwner.getUsername()
+                                )
+                );
+                return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            }
         } else {
             return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
         }
